Deal with inverted sliders storing the wrong value
Charlie, I'm guessing there's a better way to do this, but this appears to work. I *think* the issue is that if you leave the positions screen and then go back in to it then knobsInverted gets reset to false but knobs themselves don't get reset so it gets out of sync. But I could be not quite understanding.