Added reverse to Categorized Item Lists so editing state matches viewing state
Note that this means when you add a new category in the UI at the top of the list, it will actually be added at the bottom.
This isn't super-ideal but the real issue is not being able to reorder so this seems like a reasonable compromise for now.

It's definitely more important that the Crisis Types be in A-Z order!